Exception handling - delete

+++ src/main/java/com/danielalmeida/springlesson/services/UserService.java
@@ -2,8 +2,11 @@ package com.danielalmeida.springlesson.services;
 
 import com.danielalmeida.springlesson.entities.User;
 import com.danielalmeida.springlesson.repositories.UserRepository;
+import com.danielalmeida.springlesson.services.exceptions.DatabaseException;
 import com.danielalmeida.springlesson.services.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Service;
 
@@ -26,7 +29,17 @@ public class UserService {
   }
 
   public void delete(Long id){
-    repository.deleteById(id);
+    try{
+      if (repository.existsById(id)){
+        repository.deleteById(id);
+      } else {
+        throw new ResourceNotFoundException(id);
+      }
+    } catch (EmptyResultDataAccessException e){
+      throw new ResourceNotFoundException(id);
+    } catch (DataIntegrityViolationException e){
+      throw new DatabaseException(e.getMessage());
+    }
   }
 
   public User update(Long id, User obj){
